refactor(updated): Add 2024 version of old code
Due to me looking at the old code and wondering wtf I was thinking, I
made a remake of it using some of the knowledge I've gained over the
years.

It now spits out valid JSON, which can be parsed using jq or other
utillities. Ontop of this, some of the edge cases I'd missed previousy
are accounted for.

Finally, it's a bit more sane. No regex, streaming style code etc.